Prepare for visualization of internal structures
Convert trace_ridges() and combine_lines() functions into generators.
The existing behavior is preserved by helper functions to run the
generators in a loop.

The generators can be run externally to allow the algorithm to be
executed step-by-step. The goal is to interactively visualize the
internal structures at arbitrary step.