Optimize line length measurement
During the trace_ridges() processing, the 'dist' members of 'dir_arr'
are set to the accumulated distance to the seed point, instead of to the
neighbor point from 'n_dir'. This avoids the multiple re-calculations,
made by the former measure_distance().

Fix an incorrect distance in the secondary (island) "seed" points and a
minor bug in discarding of short lines.