Intermediate stage to flip the longest lines
New processing stage after the trace_ridges() one. The longest lines
that ends in a 'seed' are flipped. This moves that 'seed' and causes
adjustment of the distances along all the other lines, because of the
overlaps.

This results in creation of line along the longest possible ridge route
(main ridge). This is even valid for each of the elevation islands.