Tool to visualize the points with 'NoDataValue'
New 'boundary_info.py' module to create / append OGR polygons at the DEM
boundaries and at 'NoDataValue' DEM points. This is to ease the analysis
of 'ridges.py' results.

The tool can be run with an '-a' (append) option to add the info to
already generated ridges. It can also be used as module, i.e:
  boundary_info.create_info_geometries()