Migrate to python 3.8
The major obstacle are the new DLL import restrictions. The workaround
by setting the USE_PATH_FOR_GDAL_PYTHON=YES, does not work for the
GitHub workflow as some of the directories in PATH are missing.

This is fixed by copying all the GDAL DLLs in python's "site-packages"
directory. The PATH now points there, instead of to GDAL's "bin".