Isolate identification of the boundary pixels
The boundary pixels are identified by process_neighbors(), based on a
presence of an out-of-bounds neighbor. The former 'stop_mask' argument
is now a return value.

This allows the "split line at boundaries" rule to be also applied to
some of 'NoDataValue' points, like at the gaps created by gdal_merge.py.