Flexible info-geometry selection
Extra '--calc' option to specify an expression to calculate the mask,
around which the geometries to be created. The parameter is passed to
the python's eval() built-in:
  https://docs.python.org/3/library/functions.html#eval

The 'locals' include an 'A' variable: a numpy array with the DEM
elevations, 'dtype' is float and the "NoDataValues" are replaced with
'numpy.nan'. The expression must return a same-shaped boolean array.