Add '-boundary_val' option to mark large data-gaps
The '-boundary_val' option allows specific elevation values to be
treated as boundary points. This is similar to the 'NoDataValue',
however the points next the 'boundary_val' ones are treated like if they
are at the DEM's boundary, thus a split of fake ridges is made. See
commit id 381bf0f.

This can be useful for merged DEMs, with gaps in between the individual
sources. Previously, such gaps lead to lots of fake ridges.