Isolate array type declarations, update polylines
Move the internal array 'dtype' declarations in global constants. This
is mostly to easily maintain the keep/resume support, i.e. when these
arrays are changed.

Update the format of combine_lines() result, i.e. 'polylines'. It is now
array of dictionaries and the 'line' member is a numpy.array().