Enlarge CI validation scenario
On push event, the validation workflow runs on the complete HGT-file
(1201x1201 pixels). The minimal line processed by the 3-rd stage is
decreased to 5% of the longest line, to include more ridges in result.

This validation is also executed on pull_request and on push of branches
ending in "-test".