Allow polygons around diagonal-touching masks
Extend the create_ring_by_mask() with "outside" parameter - when "True",
the generated ring also includes diagonally adjacent regions. The
"expand_mask" class is enhanced to allow 8-direction expansion, i.e.
including the diagonals. This is to properly mask-out these rings.

This functionality is is still not in use, but is intended for future
geometries.