Allow conversion to OsmAnd OBF
When generatinf OBF files, the OsmAndMapCreator includes only geometries
with valid OSM tag. To allow conversion to OBF, all geometries are
created with "natural=ridge" or "natural=valley".

The main reference result is updated, so as the workflow validation to
pass. The old one "N39E021-mod.tif.kml" is removed, as it is obsolete.