Branch identification based on the area
Introduce a new branch identification technique, to identify the graph
branch hierarchy. It is implemented in arrange_lines(). The second and
third stages are completely rewritten to utilize this.

The technique is based on the area of coverage. The branches are
processed starting from the smallest, but this processing stops at the
graph-node points. Only the last branch to reach such point proceeds.

The reference result is updated, so as the workflow validation to pass.