Trim leaf-branches, including trunks
In the arrange_lines() preparation phase, process the "leaf" segments of
all branches in parallel. These, that ends there (leaf-branches) are
trimmed from the final result. This is valid for the branches ending at
a "seed" point, so the "leaf-trunks" are also trimmed.

This technique can be extended to the main arrange_lines() phase, to
process other non-intersecting segments in parallel.

The use of neighbor_xy() is replaced by the more versatile 'mgrid_n_xy'.