Option to separate geometries in layers
New '-multi_layer' option to separate geometries in layers. The layers
are based on the area of coverage, where 0 is the entire spheroid, each
next splits it into four (similar to the web-mercator zoom levels). This
option also selects the max zoom-level where the branches are trimmed.

The dst_layer_mgr.get_layer() is called for each branch to allow
selection of a separate layer for each geometry. The new
get_zoom_level() selects the zoom-level from the coverage area.