Extra "zoom_level" geometry field, etc.
Add a "zoom_level" field to each geometry field, based on the area of
branch coverage. This is intended to be used by the mapbox studio tool
(https://studio.mapbox.com/).

The "zoom_level" is also used to select the line widths in feature
styles, to make to make the branch hierarchy more distinguishable.