Highly parallelized branch identification
In arrange_lines(), use numpy unbuffered in place 'maximum' operation to
cut the points with less coverage-area. This avoids any python-loops
during the branch identification.

When two or more points has the same greatest coverage area, a 'fork' is
create - no branch continues, but a new one is created.

The reference result is updated, so as the workflow validation to pass.